Hide Translator class decl into cc file
